+++ test/utils/StringsTest.ts
@@ -7,14 +7,6 @@ var asr = chai.assert;
 
 suite("Strings", function StringsTest() {
 
-    test("endsWith", function endsWithTest() {
-        asr.equal(Strings.endsWith("blue", "e"), true);
-        asr.equal(Strings.endsWith("blue", "ee"), false);
-        asr.equal(Strings.endsWith("strstr", "rstr"), true);
-        asr.equal(Strings.endsWith("strstr", "strs"), false);
-    });
-
-
     test("isNullOrEmpty", function isNullOrEmptyTest() {
         asr.equal(Strings.isNullOrEmpty(""), true);
         asr.equal(Strings.isNullOrEmpty(null), true);
@@ -39,6 +31,13 @@ suite("Strings", function StringsTest() {
             var res = Strings.isCharAtDigit(str, i);
             asr.equal(res, expected[i], i + ". '" + str.substr(i, 1) + "'");
         }
+
+        asr.isTrue(Strings.isCharAtDigit("123", 1));
+        asr.isTrue(Strings.isCharAtDigit("0", 0));
+        asr.isFalse(Strings.isCharAtDigit("", 0));
+        asr.isFalse(Strings.isCharAtDigit("0A", 1));
+        asr.isFalse(Strings.isCharAtDigit("0", -1));
+        asr.isFalse(Strings.isCharAtDigit("0", 1));
     });
 
 
@@ -51,8 +50,10 @@ suite("Strings", function StringsTest() {
             asr.equal(res, expected[i], i + ". '" + str.substr(i, 1) + "'");
         }
 
-        var res2 = Strings.isDigit("123");
-        asr.equal(res2, true);
+        asr.isTrue(Strings.isDigit("123"));
+        asr.isTrue(Strings.isDigit("0"));
+        asr.isFalse(Strings.isDigit(""));
+        asr.isFalse(Strings.isDigit("0A"));
     });
 
 
@@ -64,6 +65,10 @@ suite("Strings", function StringsTest() {
             var res = Strings.isCharAtUpperCase(str, i);
             asr.equal(res, expected[i], i + ". '" + str.substr(i, 1) + "'");
         }
+
+        asr.isFalse(Strings.isCharAtUpperCase("A", -1));
+        asr.isFalse(Strings.isCharAtUpperCase("A", 1));
+        asr.isFalse(Strings.isCharAtUpperCase("A", Infinity));
     });
 
 
@@ -75,6 +80,10 @@ suite("Strings", function StringsTest() {
             var res = Strings.isCharAtLowerCase(str, i);
             asr.equal(res, expected[i], i + ". '" + str.substr(i, 1) + "'");
         }
+
+        asr.isFalse(Strings.isCharAtLowerCase("a", -1));
+        asr.isFalse(Strings.isCharAtLowerCase("a", 1));
+        asr.isFalse(Strings.isCharAtLowerCase("a", Infinity));
     });
 
 
@@ -117,6 +126,7 @@ suite("Strings", function StringsTest() {
         asr.equal(Strings.padStart(1.2, 5, " "), "  1.2");
         asr.equal(Strings.padStart(1.2, 6, "-"), "---1.2");
         asr.equal(Strings.padStart(1.2, 3, "-"), "1.2");
+        asr.equal(Strings.padStart("A", 3, "a"), "aaA");
     });
 
 
@@ -124,6 +134,7 @@ suite("Strings", function StringsTest() {
         asr.equal(Strings.padEnd(1.2, 5, " "), "1.2  ");
         asr.equal(Strings.padEnd(1.2, 6, "-"), "1.2---");
         asr.equal(Strings.padEnd(1.2, 3, "-"), "1.2");
+        asr.equal(Strings.padEnd("A", 3, "a"), "Aaa");
     });
 
 
@@ -133,6 +144,9 @@ suite("Strings", function StringsTest() {
 
         var res2 = Strings.removeLeading("---", "-", false)
         asr.equal(res2, "--");
+
+        var res3 = Strings.removeLeading("AAA", "B", false)
+        asr.equal(res3, "AAA");
     });
 
 
@@ -142,12 +156,21 @@ suite("Strings", function StringsTest() {
 
         var res2 = Strings.removeTrailing("---", "-", false)
         asr.equal(res2, "--");
+
+        var res3 = Strings.removeTrailing("AAA", "B", false)
+        asr.equal(res3, "AAA");
     });
 
 
     test("replaceAll", function replaceAllTest() {
         var res1 = Strings.replaceAll("cat in the hat", "at", "ab");
         asr.equal(res1, "cab in the hab");
+
+        var res2 = Strings.replaceAll("Super", "", "-");
+        asr.equal(res2, "S-u-p-e-r");
+
+        var res3 = Strings.replaceAll("NaN", "0", "1");
+        asr.equal(res3, "NaN");
     });
 
 });