+++ utils/Objects.ts
@@ -9,14 +9,16 @@ module Objects {
     /** Get a set of property values from an object.
      * The list of property names can be provided, or if not provided,
      * all of the object's key values will be retrieved.
-     * Example: ObjectUtil.values({ alpha: "1", beta: "2", charlie: "3" })
-     * returns: ["1", "2", "3"]
+     * Example: ObjectUtil.values({ alpha: "101", beta: "22", charlie: "3" })
+     * returns: ["101", "22", "3"]
      *
      * @param obj the object to retrieve property values from
      * @param keys optional (default: Object.keys(obj)) the list of property names
      * to retrieve from the object
      * @returns the values associated with 'keys' or 'Object.keys(obj)'
      */
+    export function values<T, K extends keyof T>(obj: T, keys?: K[] | null | undefined): (T[K])[];
+    export function values<T>(obj: { [id: string]: T } | { [id: number]: T }, keys?: string[] | null | undefined): T[];
     export function values<T>(obj: { [id: string]: T } | { [id: number]: T }, keys?: string[] | null | undefined): T[] {
         if (keys != null && !Array.isArray(keys)) {
             throw new Error("incorrect usage (" + obj + ", " + keys + "), expected (Object obj, Array<String> [keys])");
@@ -69,7 +71,7 @@ module Objects {
      * @see hasMatchingProperties()
      */
     export function hasAnyNonFalseyProps<T>(obj: T | null | undefined, propNames: (keyof T & string)[]) {
-        return hasMatchingProps(obj, propNames, function anyNonFalseyPropsFunc(val) { return !!val; }, propNames != null ? 1 : 0);
+        return hasMatchingProps(obj, propNames, function (val) { return !!val; }, propNames != null ? 1 : 0);
     }
 
 
@@ -77,7 +79,7 @@ module Objects {
      * @see hasMatchingProperties()
      */
     export function hasAnyNonNullProps<T>(obj: T | null | undefined, propNames: (keyof T & string)[]) {
-        return hasMatchingProps(obj, propNames, function anyNonNullPropsFunc(val) { return val != null; }, propNames != null ? 1 : 0);
+        return hasMatchingProps(obj, propNames, function (val) { return val != null; }, propNames != null ? 1 : 0);
     }
 
 
@@ -85,7 +87,7 @@ module Objects {
      * @see hasMatchingProperties()
      */
     export function hasNonNullProps<T>(obj: T | null | undefined, propNames: (keyof T & string)[]) {
-        return hasMatchingProps(obj, propNames, function nonNullPropsFunc(val) { return val != null; }, propNames != null ? propNames.length : 0);
+        return hasMatchingProps(obj, propNames, function (val) { return val != null; }, propNames != null ? propNames.length : 0);
     }
 
 
@@ -305,29 +307,6 @@ module Objects {
     }
 
 
-    /** Get multiple properties from an object without the risk of an undefined error.
-     * Return an empty array if either the object or the list of property names are null or undefined.
-     * Example: getProps(undefined, ["alpha", "beta"])
-     * returns: []
-     * Or example: getProps({ alpha: 342, beta: "B" }, ["alpha", "beta"])
-     * returns: [342, "B"]
-     *
-     * @param obj the object to retrieve the properties from
-     * @param propertyNames the names of the object properties to retrieve
-     * @returns the properties retrieved from the object if both the object
-     * and property names are not null, else an empty array
-     */
-    export function getProps<T, K extends keyof T>(obj: T | null | undefined, propertyNames: K[]): (T[K])[] {
-        if (obj == null || propertyNames == null || !propertyNames.length) { return []; }
-        var size = propertyNames.length;
-        var res = new Array(size);
-        for (var i = 0; i < size; i++) {
-            res[i] = obj[propertyNames[i]] || null;
-        }
-        return res;
-    }
-
-
     /** Modify classChild to extend classParent via prototypal inheritance.
      * Side-effect: classChild's prototype is modified.
      * @param classChild the sub class that inherits from 'classParent''
@@ -337,7 +316,7 @@ module Objects {
      * @param deepExtend optional (default: false) if duplicate properties are found on the 'classChild' prototype that also exist on the 'classParent' prototype, then true allows
      * the last duplicate property to take precedence, false allows the first property to take precedence. Property precedence is also determined by 'allowChildToOverride'
      */
-    export function extend(classChild: { prototype: object }, classParent: { prototype: object }, allowChildToOverride: boolean, deepExtend: boolean = false): void {
+    export function extendPrototype(classChild: { prototype: object }, classParent: { prototype: object }, allowChildToOverride: boolean, deepExtend: boolean = false): void {
         if (classParent.prototype == null) {
             throw new Error(classParent + ", does not have the property '.prototype'");
         }