+++ utils/Strings.js
@@ -4,21 +4,6 @@
  */
 var Strings;
 (function (Strings) {
-    /** Check if a string ends with a specific suffix
-     * Example: endsWith("coding in javascript", "script")
-     * returns: true
-     *
-     * @param str the string to check. Null returns false
-     * @param suffix the suffix to check for. Null returns false
-     * @returns true if the string ends with the stuffix, false otherwise
-     */
-    function endsWith(str, suffix) {
-        if (str == null || suffix == null) {
-            return false;
-        }
-        return str.indexOf(suffix) === (str.length - suffix.length);
-    }
-    Strings.endsWith = endsWith;
     /** Check if a string is null or empty
      * @param str the string to check
      * @returns true if the 'str' is null or empty, false if not
@@ -33,7 +18,7 @@ var Strings;
      * whitespace characters, false otherwise
      */
     function isNullOrWhiteSpace(str) {
-        return str == null || (str.trim instanceof Function ? str.trim().length === 0 : false);
+        return str == null || str.trim().length === 0;
     }
     Strings.isNullOrWhiteSpace = isNullOrWhiteSpace;
     /** Check if a character at a specific index in a string is a digit
@@ -51,13 +36,14 @@ var Strings;
     Strings.isCharAtDigit = isCharAtDigit;
     /** Check if all characters in a string are digits
      * @param str the string to check
-     * @returns true if every character in the string is a digit [0-9], false if not
+     * @returns true if every character in the string is a digit [0-9], false if the string is empty or contains non-digits
      */
     function isDigit(str) {
-        if (str == null) {
+        var size = 0;
+        if (str == null || (size = str.length) === 0) {
             return false;
         }
-        for (var i = 0, size = str.length; i < size; i++) {
+        for (var i = 0; i < size; i++) {
             var ch = str.charCodeAt(i) - 48;
             if (ch < 0 || ch > 9) {
                 return false;
@@ -116,30 +102,34 @@ var Strings;
         return str != null && str.length > maxLen ? str.substring(0, maxLen - ellipsis.length) + ellipsis : str || "";
     }
     Strings.clamp = clamp;
-    /** Prepend padding to the 'String(value)' representation 'value' to increase it's length to 'targetLen'
+    /** Prepend padding to the 'String(value)' representation of 'value' to increase it's length to 'targetLen'
      * @param value the value to convert to a string and pad
      * @param targetLen the maximum length of the returned string
      * @param padChar an character to use as padding
      * @returns the 'value' converted to a string and padded with 'padChar' until the string is 'targetLen' long,
-     * or returns the 'value' as a string without modification if that string is longer than 'targetLen'
+     * or returns the 'value' as a string without modification if the string is longer than 'targetLen'
      */
     function padStart(value, targetLen, padChar) {
         var valStr = String(value);
         if (valStr.length >= targetLen) {
             return valStr;
         }
-        return value != null ? new Array(targetLen - valStr.length + 1).join(padChar) + valStr : valStr;
+        return value != null ? padChar.repeat(targetLen - valStr.length) + valStr : valStr;
     }
     Strings.padStart = padStart;
-    /**
-     * @see padStart()
+    /** Append padding to the 'String(value)' representation of 'value' to increase it's length to 'targetLen'
+     * @param value the value to convert to a string and pad
+     * @param targetLen the maximum length of the returned string
+     * @param padChar an character to use as padding
+     * @returns the 'value' converted to a string and padded with 'padChar' until the string is 'targetLen' long,
+     * or returns the 'value' as a string without modification if the string is longer than 'targetLen'
      */
     function padEnd(value, targetLen, padChar) {
         var valStr = String(value);
         if (valStr.length >= targetLen) {
             return valStr;
         }
-        return value != null ? valStr + new Array(targetLen - valStr.length + 1).join(padChar) : valStr;
+        return value != null ? valStr + padChar.repeat(targetLen - valStr.length) : valStr;
     }
     Strings.padEnd = padEnd;
     /** Remove 'leadingStr' string from the beginning of 'str' as many times as it appears.